+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- New option for mapping: "Resource field" allow to choose if resource's title
+  or resource's ID should be indexed (only for values that are resources)
+- New option for mapping: "Data types" allow to limit which types of data will
+  get indexed
+
 ## [0.10.0] - 2023-05-04
 
 This version is now compatible with Omeka S 4.0.0